Scaffold out birthdays repo
