+++ src/repository/ConfigRepo.ts
@@ -13,17 +13,6 @@ export type RolesConfig = {
   muted: string;
 };
 
-export type BirthdayMessage = {
-  id: number;
-  message: string;
-};
-
-export type BirthdaysConfig = {
-  schedule: string;
-  channel: string;
-  messages: BirthdayMessage[];
-};
-
 /**
  * Default config values
  */
@@ -34,12 +23,6 @@ export const rolesConfig: RolesConfig = {
 
 export const guildConfig: GuildConfig = {
   commandPrefixes: ['-'],
-};
-
-export const birthdaysConfig: BirthdaysConfig = {
-  schedule: '00 15 10 * * *',
-  channel: '',
-  messages: [{ id: 1, message: 'Happy birthday, %!' }],
 };
 
 class ConfigRepo {
